Updated READMY and renamed interpreter
